SFR-48 Add format filter
Add `format` as a filter option for the ResearchNow search API. This filters `Work` records based off the `Link` records attached to each (within the `Item`s attached to the `Instance`s). This filters based off the `media_type` field that is associated with the `Link`s. The currently valid options for this filter are:

- pdf
- epub
- html

Additionally a new `Error` class has been added: `InvalidFilterError` that can be used to raise errors when invalid filter options are received for this and other filters.