SFR-519 Improve date parsing
This improves date cleanup and implements better parsing of uncertain/fuzzy date strings. It also begins to implement a standardization on the new ISO 8601-2:2019 standard for uncertain and approximate dates. This will be generally applied as improvements are made to the date class, ensuring that all dates can be represented in a standard format