SFR-573 Add Unicode normalization
Implements unicode normalization to "Normalization Form C" (NFC) by utilizing the `normalize` method from the built-in `unicodedata` library. This is implemented as a `listener` at the Base level of the postgreSQL ORM, validating all Unicode fields, whether passed as `bytes` or `str`

In python3.8 a new `is_normalized` method will be implemented allowing for easier checking of this. However, it is not necessary for this tool at the moment.

This will not update existing values in the database, a script will be written to validate/normalize existing values