SFR-571 Move descriptive fields to instance records
This update moves the `summary` and `table_of_contents` fields to `Instance` records from `Work`s. These fields are generally edition-specific and this update allows multiple versions of each to be stored for earch work. This allows for a fuller and more accurate description of these resources, including of versions in multiple languages.