SFR-571 Initial unglue.it retriever commit
This creates a simple service that can query the unglue.it API for an ISBN number and return metadata associated with that work from the unglue.it database. At present this is limited to the summary (if it exists) that is associated with work records. Other metadata may be fetched/returned in the future through this function.