Replace jaro_winkler with pg_trgm matching
The ResearchNow current database has used the `jaro_winkler` distance to perform fuzzy matching on text strings. This was chosen because it weights its matches towards the start of strings, providing strong matching on names. However, this incurs a significant performance penalty as the PostgreSQL implementation does not use indexes for its queries. This means that the majority of the load on the database instance is from these specific queries.

This is being replaced with the `pg_trgm` extension which implements 3-gram (trigram) matching. This splits the strings into 3-grams (e.g. "Michael" into ["Mic", "ich", "cha", "ael"]) and calculates the comparisons from these arrays. This has the following advantages:

- It can be calculated from an index, greatly speading these queries
- It can be implemented through a comparison operator, removing the need to calculate against a static score
- It handles misspellings and other common text errors accurately, without sacrificing the specificity that was gained from using the `jaro_winkler` algorithm
- The results can be easily sorted to extract the top scoring match if a match wants to be returned to the first result of multiple possible matches

Overall this should maintain the ability to perform fuzzy string matching within the database while greatly increasing the speed of the overall ingest process(in testing, a rough estimate would be by 50%)