SFR-582 Fix duplicate link generation
This fixes two related bugs that potentially allow duplicate links to be associated with a single record (either `Work`, `Instance` or `Item`). These two issues are:

- Identical resources at `HTTP` and `HTTPS` addresses
- Capitalization issues

Both of these are addressed by implementing validation on the `url` column in the `links` table. Capitalization should never be used in URLs, however, as we are ingesting data from many different sources this cannot be trusted. The `HTTP/HTTPS` issue has come up repeatedly and this should prevent similar issues in the future