SFR-491 Remove Invalid Editions
This copies the functionality from the search API response to the work response that removes invalid `instances` (lacking a publication date, place or publisher). This is implemented in the `removeInvalidEditions` method that simply filters all invalid editions.

This is implemented to prevent metadata that comes from the `instances` but is displayed at the work level from being shown in detail pages when it is removed from the search results. The bug that exposed this issue was a language from an edition that is not displayed, but which still is used for the work.