Fix addition of large identifier batches
A recent edge case exposed an issue in the `instance` class where several hundred ISBN records can cause the updater to hang as the core model attempts to find summaries for each of the ISBN records. This helps resolve the issue by streamlining the process so that only new identifiers are added to the set of existing identifiers and only those new ISBNs are queried for related summaries. While this has the immediate effect of limiting the number of queries made, it should also generally improve performance by limiting the number of requests made when adding identifiers.