Add encrypted var decoder to session manager
This commit adds handling for encrypted variables using the AWS KMS service to the session manager. This allows sensitive database credentials
to be committed to git in an encrypted form and used in a continuous deployment environment.

The change adds a static method to the `SessionManager` class that checks for a base64 encoded, encrypted string and decrypts it if found. If not, it simply returns the variable as is to allow for non-sensitive variables to be set for functions without requiring encryption. For example, an API endpoint can be added in plain text, it is not sensitive information.

To support this a new set of unit tests was added, and the `boto3` requirement was added to the pip file