HOTFIX Resolve issue with race conditions
When a race condition is found, the affected records are placed at the end of the incoming Kinesis stream. However, `instance` records are being treated as `work` records, which is causing issues. This fixes the problem by forcing them to be recognized as `instance`s. This also inmplements the `redis` cache fix, ensuring that keys have a `TTL` value