SFR-589 Add Edition table
This adds an `editions` table that can represent the results of the efforts to match/merge `instance` data. This table is relatively straightforward, containing data taken from instances for easy representation but mostly existing to record the groupings of instances and track their relationships to works. Through these relationships it is possible to quickly access any relevant data for an edition