SFR-581 Resolve timezone year range issues
When querying date ranges, such as for the publication year filter, ElasticSearch checks to see if the dates fall WITHIN the provided range, including time. To address this all dates are coerced to UTC, including the queried years from the user. However, the user supplied query was setting the search range to begin and end at noon of the first and last day of the supplied year. Internally the dates are set to midnight of December 31st, which caused ElasticSearch to see that the work dates were not within the queried dates and exclude those results.

This fix sets the query date ranges properly to midnight, which should resolve the issue.