SFR-660 Add cover updater
The ingest pipeline will store cover records asynchronously so the
updater function must be able to handle these records as they will
be sent separate from the `instance` records they will be attached
to.

This will operate by looking up the temporary, remote url that is
originally set and replace it with the local url generated by the
cover storage function.