SFR-660 importRecord refactor with cover block
This commit updates the database manager function with the same abstract model implementation as the updater function, allowing for cleaner code and enforced regularity in how different record types are implemented.

As part of this, cover parsing methods have been added to the `work` and `instance` record importers. These methods check for the presence of cover links and if found hand them to the cover processing queue for storage in s3. This allows the ingest pipeline to take covers from the data harvesters and add them to the database.

This commit also extends the test coverage for this function, both for the new importer structure and some existing methods that were previously missing