HOTFIX source data
The manager previously attempted to find the source from a child record of the record being imported, ultimately from an `item` record. However, that `source` field is difficult to access, and an easier place to locate it is in the record passed from the Kinesis stream to the function. This includes a `source` value in an easily retrievable location and allows the source to be accurately set for the cover parser, which is currently the only aspect of the function which utilizes that value