SFR-682 Add HathiTrust Cover Ingest
Add a means of retrieving cover images from the HathiTrust API. Every `bib` record in HathiTrust should be associated with a cover image, which unfortunately are not directly available through the metadata import. They can be retrived through the HathiTrust Content API, which exposes the structure of their internal eBook files as a METS object.

From this object we can determine the best cover to use for a record (the actual cover or the title page, generally). This is stored as a link on the instance record for the `bib` and which will be parsed in the `db-manager` to be stored in an s3 bucket.

This also adds some test coverage for the new classes. It is assumed that most works from HathiTrust will have covers and therefore allow ResearchNow to have a good degree of coverage in terms of covers