SFR-682 Add encrypted variables for HathiTrust
This adds the `decryptEnvVar` method from a later release of the `python-lambda-boilerplate` module that this function is based on. This allows encrytped environment varialbes to be stored in the `config` for this function and be accessed in CI/CD runs through travis. This streamlines the development process and allows it to be fully automated