SFR-683 Add cover importer to db-manager
This adds a new class for importing cover objects for instance records from the cover lookup service. These are distinct from the covers that come from the data sources in that they can be associated with any instance that does not already have a cover. The objects are stored in the database as `Link` records and associated with the correct `Instance` record before being passed to the `sfr-s3-cover` function where local, permament copies are created in s3, which are then used to replace these temporary links