SFR-683 Fix fetchers
A change to how we were getting fetched identifiers had broken this function, this commit does not revert that change but corrects it. This ensures that a found identifier is properly assigned when found.

This also fixes an issue where the mimetype of covers was being assiged to the wrong field.