HOTFIX Catch errors from bad creator field
An error where a creator could not be found in a Gutenberg record was crashing the inbex process. This simply catches the error and continues the process as we've discussed it can be a valid instance for a record to have no assigned creator. If this turns into a more frequent occurence then more investigation will be warranted.