SFR-703 Improve HathiTrust source parsing
This fixes and adds to the list of internal codes used by HathiTrust to identify organizations that have contributed to their collection either by providing materials or by digitizing those volumes.

This also centralizes this list in a single variable, which cleans up the HathiRecord class and makes it easier to reference these codes.