SFR-706 Fix Item Importer
The item importer had two different errors, both affecting the case when the importer receives a record that should be passed to the updater. First, the local reference to the SQLAlchemy `Item` record was not being properly created, and a simple query was added to create this reference. While this caused an error, this does not actually have any sideffects. Second, and more critcially, the `instance_id` was being left attached to the item record, which caused the `Item` model to crash as it is not expecting this field. This is removed when inserting an item, so it is simply removed in the same way.