HOTFIX Alt Titles
The alt title parser for the update method should not have been in a set comprehension as that was masking the fact that it could return `None`. The other option for a resolution to this issue would be to have a filtered list comprehension converted to a set, but I feel that a simple for loop is preferable for the sake of clarity.