SFR-703 Improve Identifier lookups
Requests to the database layer to retrieve a record by an array of identifiers were executing a number of sequential queries that were then iterated over to extract the best match. This has been improved by marshalling all of these queries into a single request using `UNION ALL` to join them. This cuts the execution time of the operation roughly in half, and should enable other requests to more effectively make updates in the identifier tables.