SFR-703 Performance Improvements
This adds various performance and quality of life improvements for the database updater that have already been implemented on the database manager. This is primarily focused on messaging and logging, with other minor improvements. This is also intended to be coupled with the fix to the identifier that has been put in place on the ORM layer