SFR-703 Add stream message ordering
This is an attempt to prevent database deadlocks by ordering the stream of db update messages in a more logical way. This uses the work UUID as the partition key for the updater stream, which ensures that all updates for a specific work are executed sequentially, which should both avoid race conditions and ensure that we are not making simultaneous locks on the same table rows. This is a recommended solution from Microsoft and other database providers for handling concurrency.