SFR-703 Fetch instances internally
Instead of parsing the Classify response for editions and passing the resulting list to a queue for asynchronous parsing, this moves the parsing of this list to a multithreaded module within this function. This will have the effect of moving the work of merging the work and instance records from the database update function to here, instituting a better division of concerns and potentially providing a large performance boost.

This will make this function run at a significantly reduced speed, so it should have a higher memory provision, though with concurrent executions, we should see no significant slowdown in the overall processing time.