SFR-697 Parse FrontiersIn links
This adds a new set of parsers that read the root links after following all redirects from DOAB. If these links match a specific pattern they are reformatted to a standard that we know can provide a higher degree of utility for ResearchNow users. For the FrontiersIn links this means that we can supply a local EPUB copy as well as a downloadable PDF copy for all of their open source works.

This implements a plugin structure that makes it much easier to add parsers for additional sources in the future. There is also a "base" parser that handles any links not other wise parsed in a naive way.

Finally this adds internal VIAF processing to help with the overall performance of the backend application