SFR-703 Handle multi-language strings
Several data providers will send language fields as a set of semi-colon delimited strings, which were not properly being handled before. This updates the function to make a recursive call to parse these in this case and return a list of languages.

This means that resources that have relationships to languages must be updated to handle this list. Since they simply `add`ed these to another list, this is a simple change to `extend` to handle a list of arbitrary length