SFR-697 Pass filename to the pipeline
The URLs that FrontiersIn uses to provide access cannot be used to store local copies of the ePub files because they do not provide a usable filename. This can be retrieved from the file's HEADER response so this is retrieved and parsed. That filename is then used to identify the record and a place where the file can be stored in s3 (and in a way that is easily retrievable).