SFR-703 Add VIAF name type lookup parameter
This adds the option for the user to pass the `queryType`, specifying either "personal" (the default) or "corporate"  to narrow the range of possible responeses. This helps provide a more accurate response as it prevents common surnames, which are frequently company names, from being parsed as personal names. The most common place this option will be used is when requesting publisher information, as those should always (or almost always) be corporate names