SFR-703 Update only instances associated with the current work
The new flow for the backend has eliminated updates for "bare" instances that need to be matched and associated with the proper work. Because these instances come associated with the correct work this general lookup is redundant. It can be replaced with a check for existing instances with the current work and either update or insert based off the results of this check.

This check works in the same way as the current query functions, it checks and counts the number of matching identifiers and merges the new record with the existing record with the highest number of matches. If none are found it creates a new record. (`DDC` and `LCC` identifiers are excluded from this as the are of a more general nature.)

This change has two benefits. First, it should eliminate any possibility that an instance will be associated with the wrong work. We know the right work and take that for granted with this model. Second, it should greatly reduce the number of queries to the database. While these queries are relatively fast, 100ms or so, many are made. For a large work record we might make three such queries for 100 instances. All told this would add 30 seconds to the execution time for that record. By eliminating these queries we should see marked improvement in performance