SFR-698 Add ePub link parser for Springer
This creates a new parser class for Springer, which has contributed several thousand works to the DOAB project. This adds another plugin for the LinkParser class and validates the URLs passed to it to see if they can be parsed into a resolvable epub link.

The end resuld should be that all links starting with `link.springer.com` should be resolved to a local ePub file and a PDF file that can be read online. Some links are fetched as `301` or `302` redirects and those need to be followed to find the correct form of the URL that can be paresd into these components.

There is a prefix on each of these links that usually is `10.007` but which sometimes varies, and therefore must be fetched via regex along with the ISBN number that uniquely identifies each work within the Springer collection (and hypothetically globally).