HOTFIX Remove LCCN from Classify queries
This removes LCCN as an option for making Classify queries from the manager. These queries would frequently return "multi-work" responses from the Classify API that were incorrect. Either associating multiple works by the same author or even returning results for multiple completely unrelated works. While at first glance this would seem to be an issue with metadata, this does not seem to be the problem, but instead an internal issue with the Classify API itself. LCCN is not documented as a option in the public GUI options for this service, but is present in the API documentation, which makes it seem as though its an experimental or not fully supported function.