SFR-703 Fetch all editions
Previously this was only fetching the top 500 editions of a work (sorted by the number of holdings). While this is adequate in 99% of the cases we are fetching data, it does not include all data for historically very popular works (e.g. a work by Dickens or Shakespeare).

This adds a conditional check for the number of editions associated with a work record and includes additional "paged" calls to get all of the associated works. This is not limited, so it should work in all cases. The highest number of editions are associated with Don Quixote, at around 12,000. While that may pose some timeout issues with this function, everything is broken into smaller pieces for writing to the database, so there should be no downstream side effects except for a slightly increased processing time to handle the increased edition records.