SFR-703 Multi-session execution
To deal with long execution times when the ML algorithm may take longer than the 30s idle-in-timeout limit this sets each stage of interaction (fetching, deleting old and inserting new) to operate with its own session, binding the relevant objects to these as they are created. This means that long execution times do not cause errors. (As long as the function is actively interacting with the database it will not timeout and can handle arbitrarily large records.)