SFR-774 Add cover resizer
This adds the cover resizer for the cover storage function, this re-sizes all the covers recieved to have dimensions of at most 300x400, retaining the original aspect ratio of the cover (which can be quite odd at times). This ensures that we can provide a standard output for the front end app.