SFR-774 Add mimetype to s3 storage call
By default s3 stores files as `application/octect-stream`, which forces browsers to trigger downloads of the files. Since we want them to display, we have to set these to the appropriate image mimetypes. This uses the internal python library to determine these values and passes them to s3 where they properly set this value