SFR-614 Order Editions/Instances
This returns the editions/instances returned with a search query (depending on the parameters provided) sorted in a way that matches the "parent" sort of the works as exectuted in ElasticSearch.

This is achieved by sorting the `inner_hit` records returned by ElasticSearch with each work (representing matching records) and using the found order as a clause in the Postgres query, enforcing that the order of the supplied array of `id` values is respected and represents the order of the returned rows. This ensures that first/last published versions of a work are what appear in the search results and other behavior.