SFR-667 Escape Lucene syntax characters
ElasticSearch uses the Lucene search syntax to provide a high degree of control over user queries. However the majority of people are not familiar with this and passing these characters in the wrong context can lead to 500 errors from the ES cluster. This provides a regex parser that detects and escapes these characters, with the exception of `AND`, `OR`, `*` and `"` which are all used commonly in boolean search and which do not cause search issues. This should allow users to run the queries they want without restricting their ability to control the search from the search forms.