SFR-664 Fix language counts
The function that generates the language aggregation part of the query was a static method, preventing it from accessing the instance variables it needs to accurately create these queries. Changing this, and altering the functionality of the `show_all` filter ensures that this presents accurate information to the user.

To ensure consistency this change has also been reflected in the language utility endpoint.