SFR-663 Add true edition count
Previously the edition count being displayed to the user was the number of editions that matched the current filter criteria. This is misleading since it seems to indicate the total number of editions available to view. This counts the number of editions (or instances, if we are in that context) and includes it in the `edition_count` field.