SFR-807 Add MDPI parser class
This adds a new class for parsing MDPI links. The conversion from the basic HTML page to a downloadable PDF link is simple and can be done with a simple `replace` call, which returns a single link. It would be good to provide a "Read Online" link for these books, but as our access is currently blocked this allows us to give users easy access.