HOTFIX Missing links
If an `Item` record is missing links it is an error and should be removed from the response object. This checks for the presence of the `links` array on these objects and `pop`s them from the array if none is found. This handles occasional errors from HathiTrust records where all links were found to be inaccessible, which should be removed at the time, but that cannot be 100% guaranteed. (Errors can cause an exit before a rollback can take place)

This should seamlessly handle these errors, making them invisible to users and preventing any `500`s from being thrown