SFR-692 Add Explicit fields to the keyword search
This adds more specificity to the keyword search option, restricting the queried fields to the ones that are explicity defined in the API. These were decided upon as the fields that were the most likely to be considered to be part of a keyword search by a user.

These can be updated at any time, and should provide the side effect of improving the accuracy of the search relevancy calculation. It may also have an effect on the speed of keyword searches, but small enough that it won't be noticable to the user.