SFR-814 Add parser for OpenEditions
OpenEditions is a major contributor of, mainly French, books to the DOAB project. These come from a range of publishers, but are all made available through a standard UI. These books have different levels of access from an open access page on the OE website, to a downloadable epub file. There is no API for querying for the existence of these different versions, so the HTML of the page that represents the book record must be parsed.

This is done so using the `BeatifulSoup` parser and extracts the relevant links. If epubs are found they are passed to that processing pipeline.