SFR-825 Add new fields to edition records
This adds fields that end users want to see in edition records. These fields already exist on editions in the database model but were not being populated through the clustering function. This simply passes them through, where they are picked up and indexed by the `ClusterManager`. When multiple values are found for a single edition the last one seen will be added as we have no metadata on the relative quality for the entries in these fields.

These fields are:

- Volume
- Table of Contents
- Extent
- Summary