SFR-833 Add travis integration
This adds a basic travis integration to run tests against all components of the pipeline. This currently includes no variables or build steps, which shouldn't prevent tests from passing, but any issues will be troubleshot. This is also to benchmark the build time of the pipeline to see if it is possibly long enough to interfere with NYPL's other travis jobs. I do not believe that it will be.