SFR-845 Index Works in multilanguage format
This updates the ElasticSearch index to use the per-field multilanguage indexing strategy. This means that the languages are detected using the `polyglot` library and then are both indexed in a language-specific field as well as a default field that indexes everything. This then boosts search results when terms are found in a specific language index.

In addition this now connects to the ElasticSearch endpoint via SSL to enable connections in the production environment.