SFR-877 Create Publisher Reader
This adds a framework for a metadata reader for collections of eBooks from smaller projects that have less complex metadata structures than the major projects. This enables new readers to be written and included in this framework, expanding the number of sources that can be fed into this pipeline.

This is configured as a lambda function similar to the other readers but is intended to be run less frequently than the others. Individual component readers can be activated/deactivated as necessary for greater flexibility.

At present this incorporates a single reader: for the MET Exhibition Catalog eBook collection. This is backed by an API, which makes it much easier to gather the necessary metadata.