SFR-877 Add namespace to MET identifiers
Because the local MET identifiers are being stored in the `generic` table and they are similar to other identifiers (importantly NYPL's `bibnos`) they must be distinguished from other sources. By appending `met.` to each we can easily distinguish them and denormalize them to their original forms if needed. It also keeps with the standard HathiTrust format for identifiers which should make them easily recognizable for users of this project.