SFR-877 Add VIAF/LCNAF lookup to MET importer
This adds VIAF lookup for the MET catalogs using the SFR lookup function. Corporate/Personal lookup is enabled.

This also splits the MET's `publis` field, which equates to a MARC 260 field, into publication place and publisher. This gives better data and allows for better VIAF lookups for these. It will also help with edition clustering on the other side.