SFR-877 Fetch ID and links from parent
This improves the identification and link parsing by checking for the existence of parent records, which connote an entire book/record instead of a single page.

It also switches the Read Online link to point to the catalog record as this is the only page that can be displayed in a frame through SFR. This is not an ideal solution but does provide readers with access to the materials.