HOTFIX Resolve NONE identifier bug in OCLC Lookup
The OCLC Lookup function has a special class for parsing the `856` external link MARC field. However, when it found a non-hathi link it would attempt to assign it a non-existent HathiTrust identifier. This caused these records to not be imported into the database.

This resolves the issue by ensuring that the parsed identifier is present and adding the proper source (even if it is `generic` for InternetArchive identifiers. This should resolve the issue and give us better control over the type of identifiers imported.