HOTFIX Fix bug in error handling in date module
The date module has a condition for handling ambigious YYYY-MM / YYYY-YY date values. It attempts to resolve them as year-month values but if the month value falls outside of the 1-12 range a special error is thrown. Because of a bug within the module being used to raise this error a secondary TypeError is thrown, which was not caught until a high level and which blocked the import of these records.

This udpate catches this error to ensure that they are properly parsed, including adding a new test case to prevent any regressions in this component.