SFR-891 Catch missing fields
Not all fields are returned by the InternetArchive API, this adds a few try/catch blocks to prevent unexpected errors. When these fields are not present they can simply be skipped, they never invalidate a record or make it inappropriate for importing. All the fields that would do so are mandatory.