SFR-906 Add Edition detail endpoint
Add a new endpoint that accepts an edition row id and returns a response detailing the edition and its component instances. This enables users to see the source data that was "clustered" into the edition.

The response includes metadata from the edition and an array of instances sorted by the number of holdings reported by OCLC, with the first item being a special case in that it is always the instance that contains the links made available through the edition.