SFR-906 Add showAll parameter to edition detail endpoint
This parameter controls which instance records are included in the edition detail endpoint. By default the endpoint returns all instances. If `showAll` is set to "false" only istances that have read online/download links are included are included. This enables filtering toggles on the detail page in the React app to function in a no-js environment