SFR-910 Add showAll param to work detail endpoint
Adding a `showAll` parameter to control which editions/instances are returned for work detail requests. By default, or if set to `true` this will return all instances/editions. If set to `false` it will only return instances/editions with attached items that can be read online/downloaded. This enables a new setting that are in the UI designs for the detail pages.