SFR-882 Fix OCLC 856 field parser for IA books
This parser was incorrectly treating the accessibilty field in the InternetArchive metadata response as a boolean value when it was a string true/false field.

This updates that method to take this into account and updates the relevant tests to be sure that they reflect the real-world use case. The method has also been renamed to make its function clearer.