SFR-858 Add Government Document Filter
This implements a filter on a new ElasticSearch document field `is_government_document`. This is a simple filter implemented in the API and can be included in the front end as a checkbox.

This also adds the field to the ES model in the clustering function, which will ensure that the field is populated for all newly indexed documents.