SFR-939 Add ISO_3 language code field
To support SCC (and potentially other applications), the API needs to be able to filter results based on ISO_3 language codes. Previously it only supported full language names, which is better for individual users but far less useful for integrations with other services.

This is done by adding a new condition to the language filter and transforming it into a `bool` query with two `should` clauses. This still does not use the ISO_2 field, but this is a less frequently used field and can be easily added in the future if the demand should arrise.

Testing has been done to verify that using both the full name and ISO code for a language returns identical results.