SFR-939 Fix GET endpoint for search requests
The API has always supported both GET and POST methods for the `/search` endpoint. However, the GET method was not properly parsing query parameters that should be objects, namely `sort` and `filter` (also `queries` for multi-term requests). This was due to the params being treated as is, rather than parsed as JSON to get an object. This minor update adds that functionality and ensures that the two methods operate in the same manner. After parsing the search options they both invoke the same class to execute the search, so with this fix that should be accomplished.