SFR-931 Remove Download links from Google-digitized books
The download option for Google digitized does not work offsite for NYPL because we cannot provide a partner login to HathiTrust. This may be possible in the future, but for the time being we need to remove these links. This simply prevents the download link from being generated for Google-digitized items.